Refactor ledger files
