+++ dpc/documentation/Mining_Process.md
