+++ dpc/src/ledger/memory_pool.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use snarkvm_dpc::prelude::*;
+use crate::prelude::*;
 
 use anyhow::{anyhow, Result};
 use std::collections::{HashMap, HashSet};
