+++ ledger/src/lib.rs
@@ -65,12 +65,6 @@ pub use errors::*;
 
 pub mod ledger;
 
-pub mod memory_pool;
-pub use memory_pool::*;
-
-pub mod state;
-pub use state::*;
-
 pub mod traits;
 pub use traits::*;
